# sus-nexus/contracts/openapi/core-municipal.yaml
openapi: 3.1.0
info:
  title: SUS Nexus — Core Municipal API
  version: 1.0.0
  description: |
    API interna do core municipal (Fase 1). Contract-first: o backend implementa e o frontend gera cliente a partir deste arquivo.
    Autenticação Bearer JWT (Keycloak). Erros em RFC 9457 (application/problem+json). Paginação por cursor.
servers:
  - url: http://localhost:8080
    description: dev local
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/citizens/{citizenId}:
    get:
      tags: [citizens]
      summary: Detalhe do cidadão (dados por perfil; identificadores mascarados por padrão)
      operationId: getCitizen
      parameters:
        - $ref: '#/components/parameters/citizenId'
        - $ref: '#/components/parameters/purpose'
      responses:
        '200':
          description: OK
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CitizenDetail' }
        default: { $ref: '#/components/responses/Problem' }
...
  /api/v1/citizens/{citizenId}/timeline:
    get:
      tags: [timeline]
      summary: Linha do tempo do cidadão (filtrada por perfil/finalidade no servidor)
      operationId: getTimeline
      parameters:
        - $ref: '#/components/parameters/citizenId'
        - $ref: '#/components/parameters/purpose'
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
        - name: domain
          in: query
          schema:
            type: array
            items: { $ref: '#/components/schemas/Domain' }
          style: form
          explode: false
        - name: cnes
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/TimelineEvent' }
                  next_cursor: { type: [string, 'null'] }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/citizens/{citizenId}/summary:
    get:
      tags: [timeline]
      summary: Resumo operacional (JOR-008) — sem dado clínico além do mínimo
      operationId: getCitizenSummary
      parameters:
        - $ref: '#/components/parameters/citizenId'
        - $ref: '#/components/parameters/purpose'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CitizenOperationalSummary' }
        default: { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    cursor:
      name: cursor
      in: query
      schema: { type: string }
    limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    purpose:
      name: X-Purpose-Of-Use
      in: header
      required: true
      description: Finalidade declarada do acesso (LGPD). Avaliada pela política OPA e registrada no access_log.
      schema: { $ref: '#/components/schemas/Purpose' }
...
    citizenId:
      name: citizenId
      in: path
      required: true
      schema: { type: string, pattern: '^cit_[0-9A-HJKMNP-TV-Z]{26}$' }
...
  responses:
    Problem:
      description: Erro (RFC 9457)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        correlation_id: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
    Purpose:
      type: string
      enum: [care_coordination, regulation, scheduling, identity_management, production_audit, public_health_surveillance, management_analytics, integration_operations, security_audit]
    Domain:
      type: string
      enum: [identity, aps, schedule, regulation, exam, hospital, careplan, task, production, communication]
    RegistrationState:
      type: string
      enum: [validated, divergent, incomplete, duplicate, pending]
    IdentifierSystem:
      type: string
      enum: [CNS, CPF, PEC, SISREG, ESUS_REGULACAO, HIS, AIH, APAC, LOCAL]
    MaskedIdentifier:
      type: object
      required: [id, system, value_masked, status, source_system]
      properties:
        id: { type: string }
        system: { $ref: '#/components/schemas/IdentifierSystem' }
        value_masked: { type: string }
        status: { type: string, enum: [active, deprecated, invalid] }
        source_system: { type: string }
        valid_from: { type: string, format: date-time }
    Provenance:
      type: object
      required: [source_system, received_at]
      properties:
        source_system: { type: string }
        source_record_id: { type: string }
        received_at: { type: string, format: date-time }
        confidence: { type: number, minimum: 0, maximum: 1 }
    CitizenSummary:
      type: object
      required: [id, display_name, registration_state, identifiers]
      properties:
        id: { type: string }
        display_name: { type: string, description: Nome social quando existir }
        birthdate: { type: string, format: date }
        sex: { type: string, enum: [female, male, unknown] }
        mother_name_masked: { type: string }
        registration_state: { $ref: '#/components/schemas/RegistrationState' }
        identifiers:
          type: array
          items: { $ref: '#/components/schemas/MaskedIdentifier' }
        health_unit_cnes: { type: string }
        team_ine: { type: string }
        microarea: { type: string }
        identity_confidence: { type: string, enum: [confirmed, probable, pending, divergent] }
    CitizenDetail:
      allOf:
        - $ref: '#/components/schemas/CitizenSummary'
        - type: object
          required: [version]
          properties:
            version: { type: integer }
            legal_name: { type: string }
            social_name: { type: string }
            mother_name: { type: string }
            address:
              type: object
              properties:
                street: { type: string }
                number: { type: string }
                complement: { type: string }
                district: { type: string }
                city_ibge: { type: string }
                postal_code: { type: string }
            contacts:
              type: array
              items:
                type: object
                properties:
                  kind: { type: string, enum: [phone, mobile, email] }
                  value_masked: { type: string }
                  preferred: { type: boolean }
            attribute_provenance:
              type: object
              additionalProperties: { $ref: '#/components/schemas/Provenance' }
            data_quality_issues:
              type: array
              items:
                type: object
                properties:
                  rule: { type: string }
                  field: { type: string }
                  message: { type: string }
...
    TimelineEvent:
      type: object
      required: [id, citizen_id, domain, event_type, occurred_at, recorded_at, source_system, status]
      properties:
        id: { type: string }
        citizen_id: { type: string }
        domain: { $ref: '#/components/schemas/Domain' }
        event_type: { type: string }
        occurred_at: { type: string, format: date-time }
        recorded_at: { type: string, format: date-time }
        source_system: { type: string }
        cnes: { type: string }
        health_unit_name: { type: string }
        professional_ref: { type: string }
        status: { type: string }
        confidence: { type: string, enum: [confirmed, pending, divergent, unsynced] }
        sensitivity: { type: string, enum: [public, internal, restricted, highly_restricted] }
        summary: { type: string }
        detail_ref: { type: string }
        correlation_chain:
          type: array
          items: { type: string }
    CitizenOperationalSummary:
      type: object
      required: [citizen_id]
      properties:
        citizen_id: { type: string }
        last_aps_encounter_at: { type: string, format: date-time }
        next_appointment_at: { type: string, format: date-time }
        open_tasks: { type: integer }
        open_regulation_requests: { type: integer }
        pending_exams: { type: integer }
        last_hospital_discharge_at: { type: string, format: date-time }
        care_lines: { type: array, items: { type: string } }
        care_gaps: { type: integer }
        contact_valid: { type: boolean }
...